view diem

+++ src/main/java/com/pj3/Project3/model/diem.java
@@ -12,25 +12,25 @@ public class diem {
     private Long maDiem;
 
     @Column(name = "diemLyThuyet")
-    private int diemLyThuyet;
+    private Float diemLyThuyet;
 
     @Column(name = "diemLyThuyet1")
-    private int diemLyThuyet1;
+    private Float diemLyThuyet1;
 
     @Column(name = "diemThucHanh")
-    private int diemThucHanh;
+    private Float diemThucHanh;
 
     @Column(name = "diemThucHanh1")
-    private int diemThucHanh1;
+    private Float diemThucHanh1;
 
     @Column(name = "trangThai")
-    private int trangThai;
+    private Integer trangThai;
 
     @Column(name = "diemTb")
-    private int diemTb;
+    private Float diemTb;
 
     @Column(name = "ketQua")
-    private int ketQua;
+    private Integer ketQua;
 
     @ManyToOne()
     @JoinColumn(name="maHs")
@@ -43,7 +43,7 @@ public class diem {
     public diem() {
     }
 
-    public diem(int diemLyThuyet, int diemLyThuyet1, int diemThucHanh, int diemThucHanh1, int trangThai, int diemTb, int ketQua, com.pj3.Project3.model.hocSinh hocSinh, com.pj3.Project3.model.giangDay giangDay) {
+    public diem(Float diemLyThuyet, Float diemLyThuyet1, Float diemThucHanh, Float diemThucHanh1, Integer trangThai, Float diemTb, Integer ketQua, com.pj3.Project3.model.hocSinh hocSinh, com.pj3.Project3.model.giangDay giangDay) {
         this.diemLyThuyet = diemLyThuyet;
         this.diemLyThuyet1 = diemLyThuyet1;
         this.diemThucHanh = diemThucHanh;
@@ -63,59 +63,59 @@ public class diem {
         this.maDiem = maDiem;
     }
 
-    public int getDiemLyThuyet() {
+    public Float getDiemLyThuyet() {
         return diemLyThuyet;
     }
 
-    public void setDiemLyThuyet(int diemLyThuyet) {
+    public void setDiemLyThuyet(Float diemLyThuyet) {
         this.diemLyThuyet = diemLyThuyet;
     }
 
-    public int getDiemLyThuyet1() {
+    public Float getDiemLyThuyet1() {
         return diemLyThuyet1;
     }
 
-    public void setDiemLyThuyet1(int diemLyThuyet1) {
+    public void setDiemLyThuyet1(Float diemLyThuyet1) {
         this.diemLyThuyet1 = diemLyThuyet1;
     }
 
-    public int getDiemThucHanh() {
+    public Float getDiemThucHanh() {
         return diemThucHanh;
     }
 
-    public void setDiemThucHanh(int diemThucHanh) {
+    public void setDiemThucHanh(Float diemThucHanh) {
         this.diemThucHanh = diemThucHanh;
     }
 
-    public int getDiemThucHanh1() {
+    public Float getDiemThucHanh1() {
         return diemThucHanh1;
     }
 
-    public void setDiemThucHanh1(int diemThucHanh1) {
+    public void setDiemThucHanh1(Float diemThucHanh1) {
         this.diemThucHanh1 = diemThucHanh1;
     }
 
-    public int getTrangThai() {
+    public Integer getTrangThai() {
         return trangThai;
     }
 
-    public void setTrangThai(int trangThai) {
+    public void setTrangThai(Integer trangThai) {
         this.trangThai = trangThai;
     }
 
-    public int getDiemTb() {
+    public Float getDiemTb() {
         return diemTb;
     }
 
-    public void setDiemTb(int diemTb) {
+    public void setDiemTb(Float diemTb) {
         this.diemTb = diemTb;
     }
 
-    public int getKetQua() {
+    public Integer getKetQua() {
         return ketQua;
     }
 
-    public void setKetQua(int ketQua) {
+    public void setKetQua(Integer ketQua) {
         this.ketQua = ketQua;
     }
